requirements.txt
requests~=2.28.2
Flask~=2.2.2
Gunicorn~=20.1.0
bcrypt~=4.0.1
mysql-connector-python~=8.0.32
PyJWT~=2.6.0
boto3~=1.26.74